Move old monster index classes to old_monster_index.py
