+++ dadguide/old_monster_index.py
@@ -1,18 +1,17 @@
 import difflib
 
-from collections import defaultdict
-
 from redbot.core.utils import AsyncIter
 from discord.utils import find as find_first
 import tsutils
+
+from collections import defaultdict
 
 from .models.monster_model import MonsterModel
 from .models.enum_types import Attribute
 from .models.enum_types import EvoType
 from .models.enum_types import InternalEvoType
-from .database_manager import PotentialMatches
-from .database_manager import NamedMonster
-from .database_manager import NamedMonsterGroup
+from dadguide.models.enum_types import Attribute, MonsterType
+from dadguide.models.monster_model import MonsterModel
 from .database_context import DbContext
 
 
@@ -404,3 +403,191 @@ class MonsterIndex(tsutils.aobject):
 
     def pickBestMonster(self, named_monster_list):
         return max(named_monster_list, key=lambda x: (not x.is_low_priority, x.rarity, x.monster_no_na))
+
+
+class PotentialMatches(object):
+    def __init__(self):
+        self.match_list = set()
+
+    def add(self, m):
+        self.match_list.add(m)
+
+    def update(self, monster_list):
+        self.match_list.update(monster_list)
+
+    def length(self):
+        return len(self.match_list)
+
+    def remove_potential_matches_without_all_prefixes(self, query_prefixes):
+        to_remove = set()
+        for m in self.match_list:
+            for prefix in query_prefixes:
+                if prefix not in m.prefixes:
+                    to_remove.add(m)
+                    break
+        self.match_list.difference_update(to_remove)
+
+    def get_monsters_from_potential_pantheon_match(self, pantheon, pantheon_nick_to_name, pantheons):
+        full_name = pantheon_nick_to_name[pantheon]
+        self.update(pantheons[full_name])
+
+    def pick_best_monster(self):
+        return max(self.match_list, key=lambda x: (not x.is_low_priority, x.rarity, x.monster_no_na))
+
+
+class NamedMonsterGroup(object):
+    def __init__(self, evolution_tree: list, basename_overrides: list):
+        self.is_low_priority = (
+                        self._is_low_priority_monster(evolution_tree[0])
+                        or self._is_low_priority_group(evolution_tree))
+
+        base_monster = evolution_tree[0]
+        self.group_size = len(evolution_tree)
+        self.base_monster_no = base_monster.monster_id
+        self.base_monster_no_na = base_monster.monster_no_na
+
+        self.monster_no_to_basename = {
+            m.monster_id: self._compute_monster_basename(m) for m in evolution_tree
+        }
+
+        self.computed_basename = self._compute_group_basename(evolution_tree)
+        self.computed_basenames = set([self.computed_basename])
+        if '-' in self.computed_basename:
+            self.computed_basenames.add(self.computed_basename.replace('-', ' '))
+
+        self.basenames = basename_overrides or self.computed_basenames
+
+    def _compute_monster_basename(self, m: MonsterModel):
+        basename = m.name_en.lower()
+        if ',' in basename:
+            name_parts = basename.split(',')
+            if name_parts[1].strip().startswith('the '):
+                # handle names like 'xxx, the yyy' where xxx is the name
+                basename = name_parts[0]
+            else:
+                # otherwise, grab the chunk after the last comma
+                basename = name_parts[-1]
+
+        for x in ['awoken', 'reincarnated']:
+            if basename.startswith(x):
+                basename = basename.replace(x, '')
+
+        # Fix for DC collab garbage
+        basename = basename.replace('(comics)', '')
+        basename = basename.replace('(film)', '')
+
+        return basename.strip()
+
+    def _compute_group_basename(self, monsters):
+        """Computes the basename for a group of monsters.
+
+        Prefer a basename with the largest count across the group. If all the
+        groups have equal size, prefer the lowest monster number basename.
+        This monster in general has better names, particularly when all the
+        names are unique, e.g. for male/female hunters."""
+
+        def count_and_id():
+            return [0, 0]
+
+        basename_to_info = defaultdict(count_and_id)
+
+        for m in monsters:
+            basename = self.monster_no_to_basename[m.monster_id]
+            entry = basename_to_info[basename]
+            entry[0] += 1
+            entry[1] = max(entry[1], m.monster_id)
+
+        entries = [[count_id[0], -1 * count_id[1], bn] for bn, count_id in basename_to_info.items()]
+        return max(entries)[2]
+
+    def _is_low_priority_monster(self, m: MonsterModel):
+        lp_types = [MonsterType.Evolve, MonsterType.Enhance, MonsterType.Awoken, MonsterType.Vendor]
+        lp_substrings = ['tamadra']
+        lp_min_rarity = 2
+        name = m.name_en.lower()
+
+        failed_type = m.type1 in lp_types
+        failed_ss = any([x in name for x in lp_substrings])
+        failed_rarity = m.rarity < lp_min_rarity
+        failed_chibi = name == m.name_en and m.name_en != m.name_ja
+        failed_equip = m.is_equip
+        return failed_type or failed_ss or failed_rarity or failed_chibi or failed_equip
+
+    def _is_low_priority_group(self, mg: list):
+        lp_grp_min_rarity = 5
+        max_rarity = max(m.rarity for m in mg)
+        failed_max_rarity = max_rarity < lp_grp_min_rarity
+        return failed_max_rarity
+
+
+class NamedMonster(object):
+    def __init__(self, monster: MonsterModel, monster_group: NamedMonsterGroup, prefixes: set, extra_nicknames: set, db_context=None):
+        # Must not hold onto monster or monster_group!
+
+        # Hold on to the IDs instead
+        self.monster_id = monster.monster_id
+        self.monster_no_na = monster.monster_no_na
+        self.monster_no_jp = monster.monster_no_jp
+
+        # ID of the root of the tree for this monster
+        self.base_monster_no = monster_group.base_monster_no
+        self.base_monster_no_na = monster_group.base_monster_no_na
+
+        # This stuff is important for nickname generation
+        self.group_basenames = monster_group.basenames
+        self.prefixes = prefixes
+
+        # Pantheon
+        series = db_context.graph.get_monster(monster.monster_no).series
+        self.series = series.name if series else None
+
+        # Data used to determine how to rank the nicknames
+        self.is_low_priority = monster_group.is_low_priority or monster.is_equip
+        self.group_size = monster_group.group_size
+        self.rarity = monster.rarity
+
+        # Used in fallback searches
+        self.name_en = monster.name_en
+        self.name_ja = monster.name_ja
+
+        # These are just extra metadata
+        self.monster_basename = monster_group.monster_no_to_basename[self.monster_id]
+        self.group_computed_basename = monster_group.computed_basename
+        self.extra_nicknames = extra_nicknames
+
+        # Compute any extra prefixes
+        if self.monster_basename in ('ana', 'ace'):
+            self.prefixes.add(self.monster_basename)
+
+        # Compute extra basenames by checking for two-word basenames and using the second half
+        self.two_word_basenames = set()
+        for basename in self.group_basenames:
+            basename_words = basename.split(' ')
+            if len(basename_words) == 2:
+                self.two_word_basenames.add(basename_words[1])
+
+        # The primary result nicknames
+        self.final_nicknames = set()
+        # Set the configured override nicknames
+        self.final_nicknames.update(self.extra_nicknames)
+        # Set the roma subname for JP monsters
+        if monster.roma_subname:
+            self.final_nicknames.add(monster.roma_subname)
+
+        # For each basename, add nicknames
+        for basename in self.group_basenames:
+            # Add the basename directly
+            self.final_nicknames.add(basename)
+            # Add the prefix plus basename, and the prefix with a space between basename
+            for prefix in self.prefixes:
+                self.final_nicknames.add(prefix + basename)
+                self.final_nicknames.add(prefix + ' ' + basename)
+
+        self.final_two_word_nicknames = set()
+        # Slightly different process for two-word basenames. Does this make sense? Who knows.
+        for basename in self.two_word_basenames:
+            self.final_two_word_nicknames.add(basename)
+            # Add the prefix plus basename, and the prefix with a space between basename
+            for prefix in self.prefixes:
+                self.final_two_word_nicknames.add(prefix + basename)
+                self.final_two_word_nicknames.add(prefix + ' ' + basename)
